--- libs/namelog/UserHistory.json
@@ -1,1 +1,0 @@
-[{"user": "Gogurk", "time": "Fri Apr 21 14:44:12 2023"}]
