+++ libs/resources/resources/logs/UserHistory.json
@@ -1,0 +1,1 @@
+[{"user": "Gogurk", "time": "Wed Apr 26 10:25:34 2023"}][{"user": "Gogurk", "time": "Wed Apr 26 10:39:45 2023"}][{"user": "Gogurk", "time": "Wed Apr 26 10:42:05 2023"}][{"user": "Gogurk", "time": "Fri Apr 28 10:36:17 2023"}]
